Work on center list

+++ database/migrations/2024_08_19_081615_create_exams_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('name')->unique();
             $table->date('reg_last_date');
             $table->date('reg_final_date');
-            $table->decimal('registration_fee', 8);
-            $table->decimal('late_fee', 8);
+            $table->unsignedSmallInteger('registration_fee');
+            $table->unsignedSmallInteger('late_fee');
             $table->timestamps();
         });
     }
